Add CentOS to CI (#45)

+++ .ci/unix/dump-vm-specs.sh
@@ -15,7 +15,11 @@ if [ "$(uname)" == "Darwin" ]; then
     sudo scutil --get LocalHostName || true
 elif [ "$(uname)" == "Linux" ]; then
     if [ "$(which lsb_release)" == "" ]; then
-        sudo apt install -y lsb-release
+        if [ -f /etc/redhat-release ]; then
+            sudo yum install -y redhat-lsb-core
+        else
+            sudo apt install -y lsb-release
+        fi
     fi
     lsb_release -a
     free -m
